Fix test discovery issues and disable Auth0 legacy tests
- Disable Auth0 related test files (test_auth0.py, test_settings.py) as Auth0 is no longer used
- Disable pytest-specific test file (test_models_pytest.py) that conflicts with Django test runner
- Fix CI test discovery by specifying explicit test patterns and modules
- Add --pattern="test_*.py" to prevent loading non-test modules as tests
- Clean up test module loading to avoid unittest.loader._FailedTest errors
