+++ .github/workflows/ci-cd-pipeline.yml
@@ -98,6 +98,7 @@ jobs:
           echo "TEST_MODE=True" >> .env
           echo "DJANGO_ENV=ci" >> .env
           echo "ALLOWED_HOSTS=*" >> .env
+          echo "AUTH0_CLIENT_SECRET=test-secret" >> .env
 
       - name: Run migrations
         run: |
@@ -114,7 +115,15 @@ jobs:
           source .env
           set +a
           echo "🚀 Running Complete Test Suite"
-          poetry run python manage.py test --verbosity=2 --keepdb
+          # Run tests using test discovery but exclude problematic directories
+          poetry run python manage.py test \
+            tests \
+            apps.authentication.tests \
+            apps.quizz.tests \
+            apps.revision.tests \
+            apps.task.tests \
+            core.jobs.tests \
+            --verbosity=2 --keepdb --settings=core.settings_test --pattern="test_*.py"
         env:
           TEST_MODE: 'True'
           CI: 'True'
@@ -239,6 +248,7 @@ jobs:
           echo "TEST_MODE=True" >> .env
           echo "DJANGO_ENV=ci" >> .env
           echo "ALLOWED_HOSTS=*" >> .env
+          echo "AUTH0_CLIENT_SECRET=test-secret" >> .env
 
       - name: Run migrations
         run: |
@@ -255,8 +265,16 @@ jobs:
           source .env
           set +a
           echo "🔥 Running FULL Test Suite for Staging"
-          poetry run python manage.py test --verbosity=2 --keepdb
-          echo "🎯 Running Jobs Tests"
+          # Run specific test modules to avoid loading non-test files
+          poetry run python manage.py test \
+            apps.authentication.tests \
+            apps.quizz.tests \
+            apps.revision.tests \
+            apps.task.tests \
+            core.jobs.tests \
+            tests \
+            --verbosity=2 --keepdb --settings=core.settings_test
+          echo "🎯 Running Jobs Tests Again for Verification"
           poetry run python manage.py test core.jobs.tests -v 2 --settings=core.settings_test --keepdb
         env:
           TEST_MODE: 'True'
