+++ backend/apps/authentication/tests/test_models_pytest.py.disabled
@@ -7,7 +7,7 @@ import pytest
 import datetime
 from decimal import Decimal
 from django.db import IntegrityError
-from ..models import User, CoachProfile, Review, UserFeedback
+from backend.apps.authentication.models import User, CoachProfile, Review, UserFeedback
 
 # Tests pour le modèle User
 @pytest.mark.django_db
