+++ backend/apps/authentication/tests/test_settings.py.disabled
@@ -3,9 +3,7 @@
 Configuration spéciale pour les tests de base de données.
 Ce fichier peut être importé dans vos tests ou dans conftest.py
 """
-import environ
-
-env = environ.Env()
+import os
 # Configuration de la base de données de test
 TEST_DB = {
     'default': {
@@ -34,7 +32,7 @@ SQLITE_MEMORY_DB = {
 AUTH0_SETTINGS = {
     'AUTH0_DOMAIN': 'dev-7qe275o831ebkhbj.eu.auth0.com',
     'AUTH0_CLIENT_ID': '3e0iFOwYLe3h2QkrqrxtSu8h3gGah2ml',
-    'AUTH0_CLIENT_SECRET': env('AUTH0_CLIENT_SECRET'),
+    'AUTH0_CLIENT_SECRET': os.environ.get('AUTH0_CLIENT_SECRET', 'test-secret'),
     'AUTH0_AUDIENCE': 'https://linguify-api',
     'FRONTEND_URL': 'http://localhost:3000',
     'FRONTEND_CALLBACK_URL': 'http://localhost:3000/callback',
